Fix CI
- Fix issue described in https://github.com/dethcrypto/TypeChain/issues/371
