+++ .github/workflows/coverage.yaml
@@ -20,6 +20,7 @@ jobs:
           restore-keys: |
             ${{ runner.os }}-yarn-
       - run: yarn install
+      - run: yarn typechain
       - run: yarn compile
       - run: yarn coverage
