+++ .github/workflows/lint-sol.yaml
@@ -20,4 +20,6 @@ jobs:
           restore-keys: |
             ${{ runner.os }}-yarn-
       - run: yarn install
+      - run: yarn typechain
+      - run: yarn compile
       - run: yarn lint-sol
