+++ .github/workflows/lint-ts.yaml
@@ -20,5 +20,6 @@ jobs:
           restore-keys: |
             ${{ runner.os }}-yarn-
       - run: yarn install
+      - run: yarn typechain
       - run: yarn compile
       - run: yarn lint-ts
