+++ .github/workflows/test.yaml
@@ -20,5 +20,6 @@ jobs:
           restore-keys: |
             ${{ runner.os }}-yarn-
       - run: yarn
+      - run: yarn typechain
       - run: yarn compile
       - run: yarn test
