--- tasks.config.ts
@@ -1,8 +1,0 @@
-import config from './hardhat.config'
-
-import './tasks/deployMerkleDrop'
-import './tasks/deployMerkleDropTranches'
-import './tasks/seedNewAllocations'
-import './tasks/addTrancheURI'
-
-export default config
